Update code to new controller based organisation
